Genetic algo further implemented, ITSP individuals created.

+++ src/itsp/DistanceMetric.java
@@ -2,6 +2,13 @@ package itsp;
 
 public interface DistanceMetric {
 
-    public double distanceBetween(ProcessingNode n1, ProcessingNode n2);
+    /**
+     * Calculates the distance between two nodes, expressed in timesteps.
+     *
+     * @param n1
+     * @param n2
+     * @return
+     */
+    public int distanceBetween(ProcessingNode n1, ProcessingNode n2);
 
 }
